+++ apphelpers/rest/hug.py
@@ -11,11 +11,11 @@ from apphelpers.db import dbtransaction
 from apphelpers.errors.hug import BaseError, InvalidSessionError
 from apphelpers.loggers import api_logger
 from apphelpers.rest import endpoint as ep
+from apphelpers.rest.common import notify_honeybadger
 from apphelpers.sessions import SessionDBHandler
 
 if settings.get("HONEYBADGER_API_KEY"):
     from honeybadger import Honeybadger
-    from honeybadger.utils import filter_dict
 
 
 def phony(f):
@@ -36,20 +36,6 @@ def raise_not_found_on_none(f):
     return f
 
 
-def notify_honeybadger(honeybadger, error, func, args, kwargs):
-    try:
-        honeybadger.notify(
-            error,
-            context={
-                "func": func.__name__,
-                "args": args,
-                "kwargs": filter_dict(kwargs, settings.HB_PARAM_FILTERS),
-            },
-        )
-    finally:
-        pass
-
-
 def honeybadger_wrapper(hb):
     """
     wrapper that executes the function in a try/except
@@ -59,20 +45,25 @@ def honeybadger_wrapper(hb):
     def wrapper(f):
         @wraps(f)
         def f_wrapped(*args, **kw):
+            err_to_report = None
             try:
                 return f(*args, **kw)
             except BaseError as e:
                 if e.report:
-                    notify_honeybadger(
-                        honeybadger=hb, error=e, func=f, args=args, kwargs=kw
-                    )
+                    err_to_report = e
                 raise e
-
             except Exception as e:
-                notify_honeybadger(
-                    honeybadger=hb, error=e, func=f, args=args, kwargs=kw
-                )
+                err_to_report = e
                 raise e
+            finally:
+                if err_to_report:
+                    notify_honeybadger(
+                        honeybadger=hb,
+                        error=err_to_report,
+                        func=f,
+                        args=args,
+                        kwargs=kw,
+                    )
 
         return f_wrapped